refactor: rename db_seed module
