+++ app/db_seed/adapters/base_adapter.py
