+++ app/db_seed/adapters/sqlite/sqlite_order_detail_adapter.py
@@ -1,6 +1,5 @@
-from typing import List
-from app.db_populator.adapters.base_adapter import BaseRepository
-from app.db_populator.db.session import conn, cursor
+from app.db_seed.adapters.base_adapter import BaseRepository
+from app.db_seed.db.session import conn, cursor
 
 
 class SQLiteOrderDetailAdapter(BaseRepository):
